+++ docs/examples/opencensus-exporter-metrics/collector.py
@@ -17,7 +17,7 @@ This example shows how to export metrics to the OT collector.
 """
 
 from opentelemetry import metrics
-from opentelemetry.ext.opencensusexporter.metrics_exporter import (
+from opentelemetry.exporter.opencensus.metrics_exporter import (
     OpenCensusMetricsExporter,
 )
 from opentelemetry.sdk.metrics import Counter, MeterProvider
