+++ docs/examples/opencensus-exporter-tracer/collector.py
@@ -15,7 +15,7 @@
 # limitations under the License.
 
 from opentelemetry import trace
-from opentelemetry.ext.opencensusexporter.trace_exporter import (
+from opentelemetry.exporter.opencensus.trace_exporter import (
     OpenCensusSpanExporter,
 )
 from opentelemetry.sdk.trace import TracerProvider
