+++ docs/examples/opentracing/main.py
@@ -2,7 +2,7 @@
 
 from opentelemetry import trace
 from opentelemetry.ext import opentracing_shim
-from opentelemetry.ext.jaeger import JaegerSpanExporter
+from opentelemetry.exporter.jaeger import JaegerSpanExporter
 from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import SimpleExportSpanProcessor
 from rediscache import RedisCache
