+++ exporter/opentelemetry-exporter-jaeger/src/opentelemetry/exporter/jaeger/__init__.py
@@ -69,8 +69,8 @@ from thrift.protocol import TBinaryProtocol, TCompactProtocol
 from thrift.transport import THttpClient, TTransport
 
 import opentelemetry.trace as trace_api
-from opentelemetry.ext.jaeger.gen.agent import Agent as agent
-from opentelemetry.ext.jaeger.gen.jaeger import Collector as jaeger
+from opentelemetry.exporter.jaeger.gen.agent import Agent as agent
+from opentelemetry.exporter.jaeger.gen.jaeger import Collector as jaeger
 from opentelemetry.sdk.trace.export import Span, SpanExporter, SpanExportResult
 from opentelemetry.trace.status import StatusCanonicalCode
