+++ exporter/opentelemetry-exporter-opencensus/src/opentelemetry/exporter/opencensus/util.py
@@ -21,7 +21,7 @@ from google.protobuf.timestamp_pb2 import Timestamp
 from opencensus.proto.agent.common.v1 import common_pb2
 from opencensus.proto.trace.v1 import trace_pb2
 
-from opentelemetry.ext.opencensusexporter.version import (
+from opentelemetry.exporter.opencensus.version import (
     __version__ as opencensusexporter_exporter_version,
 )
 from opentelemetry.trace import SpanKind
