+++ exporter/opentelemetry-exporter-opencensus/tests/test_otcollector_metrics_exporter.py
@@ -20,7 +20,7 @@ from google.protobuf.timestamp_pb2 import Timestamp
 from opencensus.proto.metrics.v1 import metrics_pb2
 
 from opentelemetry import metrics
-from opentelemetry.ext.opencensusexporter import metrics_exporter
+from opentelemetry.exporter.opencensus import metrics_exporter
 from opentelemetry.sdk.metrics import (
     Counter,
     MeterProvider,
@@ -47,7 +47,7 @@ class TestCollectorMetricsExporter(unittest.TestCase):
     def test_constructor(self):
         mock_get_node = mock.Mock()
         patch = mock.patch(
-            "opentelemetry.ext.opencensusexporter.util.get_node",
+            "opentelemetry.exporter.opencensus.util.get_node",
             side_effect=mock_get_node,
         )
         service_name = "testServiceName"
